requirements.txt
plotly==5.9.0
streamlit-authenticator==0.2.2
Pillow==9.4.0
pandas==2.0.2